Show error message if loadInitialState fails
